Corregidos Servlets, añadidos algunos métodos en EJB, falta definir el ámbito

+++ CityPlanGit/src/bl/Valoracion.java
@@ -6,7 +6,7 @@ import dl.Evento;
 public class Valoracion {
 	private Cliente cliente;
 	private Evento evento;
-	private int totalValoracion;
+	private float totalValoracion;
 	private String respuestasValoracion;
 
 	public Valoracion() {
@@ -29,11 +29,11 @@ public class Valoracion {
 		this.evento = evento;
 	}
 
-	public int getTotalValoracion() {
+	public float getTotalValoracion() {
 		return totalValoracion;
 	}
 
-	public void setTotalValoracion(int totalValoracion) {
+	public void setTotalValoracion(float totalValoracion) {
 		this.totalValoracion = totalValoracion;
 	}
 
@@ -46,7 +46,9 @@ public class Valoracion {
 	}
 
 	// Calcular nuevo valor de la valoración:
-	public int calculaValoracion(int p1, int p2, int p3, int p4, int p5, int p6, int p7, int p8) {
+	// NO ESTÁ DEL TODO BIEN, FALTA METER UN PARÁMETRO CON LA VALORACIÓN ANTERIOR Y
+	// EL NÚMERO DE VALORACIONES EXISTENTES
+	public float calculaValoracion(int p1, int p2, int p3, int p4, int p5, int p6, int p7, int p8) {
 
 		totalValoracion = (p1 + p2 + p3 + p4 + p5 + p6 + p7 + p8) / 8;
 		return totalValoracion;
